requirements.txt
google-cloud-bigquery==3.25.0
google-cloud-storage==2.18.2
google-cloud-bigquery-datatransfer==3.15.6
tqdm==4.66.5
